+++ Programacion/ProyectoIntermodular/src/main/java/dto/Disco.java
@@ -12,9 +12,9 @@ public class Disco extends Producto {
     private final String sello;
     private final String artista;
     private final String tipo;
-    private final int asin;
+    private final long asin;
 
-    public Disco(String canciones, String sello, String artista, String tipo, int asin, int codigo, int iva, int stock, int stockMinimo, String unidadDeMedida, double precio, String nombre, String descripcion, String foto, LocalDateTime fechaCreacion, LocalDateTime fechaUltimaModificacion, Proveedor proveedor, Usuario creador, Usuario modificador, ArrayList<Categoria> categorias) {
+    public Disco(String canciones, String sello, String artista, String tipo, long asin, int codigo, int iva, int stock, int stockMinimo, String unidadDeMedida, double precio, String nombre, String descripcion, String foto, LocalDateTime fechaCreacion, LocalDateTime fechaUltimaModificacion, Proveedor proveedor, Usuario creador, Usuario modificador, ArrayList<Categoria> categorias) {
         super(codigo, iva, stock, stockMinimo, unidadDeMedida, precio, nombre, descripcion, foto, fechaCreacion, fechaUltimaModificacion, proveedor, creador, modificador, categorias);
         this.canciones = canciones;
         this.sello = sello;
@@ -23,7 +23,7 @@ public class Disco extends Producto {
         this.asin = asin;
     }
     
-    public Disco(Producto p, String canciones, String sello, String artista, String tipo, int asin) {
+    public Disco(Producto p, String canciones, String sello, String artista, String tipo, long asin) {
         this(canciones, sello, artista, tipo, asin, p.getCodigo(), p.getIva(), p.getStock(), p.getStockMinimo(), p.getUnidadDeMedida(), p.getPrecio(), p.getNombre(), p.getDescripcion(), p.getFoto(), p.getFechaCreacion(), p.getFechaUltimaModificacion(), p.getProveedor(), p.getCreador(), p.getModificador(), p.getCategorias());
     }
 
@@ -43,7 +43,7 @@ public class Disco extends Producto {
         return tipo;
     }
 
-    public int getAsin() {
+    public long getAsin() {
         return asin;
     }
 }
